+++ CHANGELOG.md
@@ -71,6 +71,8 @@ BUG FIXES:
   - providers/virtualbox: Make more internal interactions with VBoxManage
     retryable to avoid spurious VirtualBox errors. [GH-2831]
   - providers/virtualbox: Config validation catches invalid keys. [GH-2843]
+  - providers/virtualbox: Fix network adapter configuration issue if using
+    provider-specific config. [GH-2854]
   - provisioners/chef: When chowning folders, don't follow symlinks.
   - provisioners/docker: Only add SSH user to docker group if the user
     isn't already in it. [GH-2838]